Add Calculator v.2.0 with two Tab-Fragments and string values for text to this project.
Fill DbHelper with all fields to handle distances and stage of Bike Tourism (not  check it) and change add stage activity to handle it (work properly)

Handle SharedPreferences for competitionId and distanceId.

Create hidden field for distanceId in item_competition.xml and set it with old CompetitionCursorAdapter (add column distanceId to ContentProvider)

Add Id for distances to Contract file and I use it.